remove peth v2
