+++ bin/peth/Cargo.toml
@@ -17,114 +17,19 @@ futures = "0.3.30"
 dirs = "5.0.1"
 tokio = "1.40.0"
 tokio-stream = "0.1.16"
-
-
 eyre = "0.6.12"
 tracing = "0.1.40"
 tempfile = "3.12.0"
-api.workspace = true
-api-types.workspace = true
-jsonrpsee = "=0.24.4"
+jsonrpsee = "0.24"
 hex = "0.4.3"
 serde_json = "1.0.128"
 bincode = "1.3"
 time = "0.3.36"
 anyhow = "1.0.87"
-alloy = { version = "0.3.6", features = ["eip712", "consensus", "rlp", "eips"] }
-revm = { version = "=14.0.2" }
-revm-inspectors = { version = "=0.7" }
-revm-primitives = { version = "=9.0.2", features = [
-    "std",
-], default-features = false }
-alloy-chains = "=0.1.32"
-alloy-dyn-abi = "=0.8.3"
-alloy-primitives = { version = "=0.8.3", default-features = false }
-alloy-rlp = "=0.3.4"
-alloy-sol-types = "=0.8.3"
-alloy-trie = { version = "=0.5", default-features = false }
-alloy-sol-macro = { version = "=0.8.3" }
-alloy-consensus = { version = "=0.3.6", default-features = false }
-alloy-eips = { version = "=0.3.6", default-features = false }
-alloy-genesis = { version = "=0.3.6", default-features = false }
-alloy-json-rpc = { version = "=0.3.6", default-features = false }
-alloy-network = { version = "=0.3.6", default-features = false }
-alloy-network-primitives = { version = "=0.3.6", default-features = false }
-alloy-node-bindings = { version = "=0.3.6", default-features = false }
-alloy-provider = { version = "=0.3.6", features = [
-    "reqwest",
-], default-features = false }
-alloy-sol-macro-expander = { version = "=0.8.3" }
-alloy-sol-macro-input = { version = "=0.8.3" }
-alloy-pubsub = { version = "=0.3.6", default-features = false }
-alloy-rpc-client = { version = "=0.3.6", default-features = false }
-alloy-rpc-types = { version = "=0.3.6", features = [
-    "eth",
-], default-features = false }
-alloy-sol-type-parser = { version = "=0.8.3" }
-alloy-rpc-types-admin = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-anvil = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-beacon = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-debug = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-engine = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-eth = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-mev = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-trace = { version = "=0.3.6", default-features = false }
-alloy-rpc-types-txpool = { version = "=0.3.6", default-features = false }
-alloy-serde = { version = "=0.3.6", default-features = false }
-alloy-signer = { version = "=0.3.6", default-features = false }
-alloy-signer-local = { version = "=0.3.6", default-features = false }
-alloy-transport = { version = "=0.3.6" }
-alloy-transport-http = { version = "=0.3.6", features = [
-    "reqwest-rustls-tls",
-], default-features = false }
-alloy-transport-ipc = { version = "=0.3.6", default-features = false }
-alloy-transport-ws = { version = "=0.3.6", default-features = false }
-
-# op
-op-alloy-rpc-types = "0.2.12"
-op-alloy-rpc-types-engine = "0.2.12"
-op-alloy-network = "0.2.12"
-op-alloy-consensus = "0.2.12"
-serde = { version = "1.0.128", features = ["derive"] }
-web3 = "0.19.0"
-
-reth = { git = "https://github.com/Galxe/grevm-reth", package = "reth", branch = "v1.0.7-gravity" }
-reth-node-builder = { git = "https://github.com/Galxe/grevm-reth", package = "reth-node-builder", branch = "v1.0.7-gravity" }
-reth-node-ethereum = { git = "https://github.com/Galxe/grevm-reth", package = "reth-node-ethereum", branch = "v1.0.7-gravity" }
-reth-node-core = { git = "https://github.com/Galxe/grevm-reth", package = "reth-node-core", branch = "v1.0.7-gravity" }
-reth-provider = { git = "https://github.com/Galxe/grevm-reth", package = "reth-provider", branch = "v1.0.7-gravity" }
-reth-db = { git = "https://github.com/Galxe/grevm-reth", package = "reth-db", branch = "v1.0.7-gravity" }
-reth-tasks = { git = "https://github.com/Galxe/grevm-reth", package = "reth-tasks", branch = "v1.0.7-gravity" }
-reth-cli-runner = { git = "https://github.com/Galxe/grevm-reth", package = "reth-cli-runner", branch = "v1.0.7-gravity" }
-reth-cli = { git = "https://github.com/Galxe/grevm-reth", package = "reth-cli", branch = "v1.0.7-gravity" }
-reth-cli-util = { git = "https://github.com/Galxe/grevm-reth", package = "reth-cli-util", branch = "v1.0.7-gravity" }
-reth-chainspec = { git = "https://github.com/Galxe/grevm-reth", package = "reth-chainspec", branch = "v1.0.7-gravity" }
-reth-tracing = { git = "https://github.com/Galxe/grevm-reth", package = "reth-tracing", branch = "v1.0.7-gravity" }
-reth-auto-seal-consensus = { git = "https://github.com/Galxe/grevm-reth", package = "reth-auto-seal-consensus", branch = "v1.0.7-gravity" }
-reth-basic-payload-builder = { git = "https://github.com/Galxe/grevm-reth", package = "reth-basic-payload-builder", branch = "v1.0.7-gravity" }
-reth-beacon-consensus = { git = "https://github.com/Galxe/grevm-reth", package = "reth-beacon-consensus", branch = "v1.0.7-gravity" }
-reth-ethereum-engine-primitives = { git = "https://github.com/Galxe/grevm-reth", package = "reth-ethereum-engine-primitives", branch = "v1.0.7-gravity" }
-reth-evm-ethereum = { git = "https://github.com/Galxe/grevm-reth", package = "reth-evm-ethereum", branch = "v1.0.7-gravity" }
-reth-network = { git = "https://github.com/Galxe/grevm-reth", package = "reth-network", branch = "v1.0.7-gravity" }
-reth-node-api = { git = "https://github.com/Galxe/grevm-reth", package = "reth-node-api", branch = "v1.0.7-gravity" }
-reth-payload-builder = { git = "https://github.com/Galxe/grevm-reth", package = "reth-payload-builder", branch = "v1.0.7-gravity" }
-reth-rpc = { git = "https://github.com/Galxe/grevm-reth", package = "reth-rpc", branch = "v1.0.7-gravity" }
-reth-rpc-api = { git = "https://github.com/Galxe/grevm-reth", package = "reth-rpc-api", branch = "v1.0.7-gravity", features = [
-    "client",
-] }
-reth-rpc-layer = { git = "https://github.com/Galxe/grevm-reth", package = "reth-rpc-layer", branch = "v1.0.7-gravity" }
-reth-rpc-types = { git = "https://github.com/Galxe/grevm-reth", package = "reth-rpc-types", branch = "v1.0.7-gravity" }
-reth-transaction-pool = { git = "https://github.com/Galxe/grevm-reth", package = "reth-transaction-pool", branch = "v1.0.7-gravity" }
-reth-consensus = { git = "https://github.com/Galxe/grevm-reth", package = "reth-consensus", branch = "v1.0.7-gravity" }
-reth-ethereum-payload-builder = { git = "https://github.com/Galxe/grevm-reth", package = "reth-ethereum-payload-builder", branch = "v1.0.7-gravity" }
-reth-primitives = { git = "https://github.com/Galxe/grevm-reth", package = "reth-primitives", branch = "v1.0.7-gravity" }
-reth-cli-commands = { git = "https://github.com/Galxe/grevm-reth", package = "reth-cli-commands", features = [
-    "dev",
-], branch = "v1.0.7-gravity" }
-reth-pipe-exec-layer-ext = { git = "https://github.com/Galxe/grevm-reth", package = "reth-pipe-exec-layer-ext", branch = "v1.0.7-gravity" }
+greth = { git = "https://github.com/Galxe/grevm-reth", branch = "main" }
 reqwest = "0.12.9"
-async-trait.workspace = true
 
-# We'd better specify the version of wasm-bindgen, since some version like '0.2.97' can't compile in Linux
-[dependencies.wasm-bindgen]
-version = "0.2.99"
+async-trait.workspace = true
+api.workspace = true
+api-types.workspace = true
+
