+++ bin/peth/src/reth_cli.rs
@@ -1,328 +1,264 @@
-use alloy_consensus::{TxEip1559, TxEip2930, TxLegacy};
-use alloy_primitives::{Address, B256};
-use alloy_rpc_types_engine::ForkchoiceState;
+use crate::ConsensusArgs;
 use api_types::account::{ExternalAccountAddress, ExternalChainId};
 use api_types::u256_define::{BlockId as ExternalBlockId, TxnHash};
-use api_types::{simple_hash, ExternalBlock, VerifiedTxn, VerifiedTxnWithAccountSeqNum};
-use jsonrpsee::core::Serialize;
-use jsonrpsee::http_client::transport::HttpBackend;
-use jsonrpsee::http_client::HttpClient;
-use reth::rpc::builder::auth::AuthServerHandle;
-use reth_ethereum_engine_primitives::{EthEngineTypes, EthPayloadAttributes};
-use reth_payload_builder::PayloadId;
-use reth_pipe_exec_layer_ext::{OrderedBlock, PipeExecLayerApi};
-use reth_primitives::{Bytes, Signature, TransactionSigned, TxKind, U256};
-use reth_rpc_api::{EngineApiClient, EngineEthApiClient};
-use reth_rpc_layer::AuthClientService;
-use reth_rpc_types::{AccessList, AccessListItem};
+use api_types::{ExecutionBlocks, ExternalBlock, VerifiedTxn, VerifiedTxnWithAccountSeqNum};
+use core::panic;
+use greth::reth::rpc::builder::auth::AuthServerHandle;
+use greth::reth_db::DatabaseEnv;
+use greth::reth_ethereum_engine_primitives::EthPayloadAttributes;
+use greth::reth_node_api::NodeTypesWithDBAdapter;
+use greth::reth_node_ethereum::EthereumNode;
+use greth::reth_pipe_exec_layer_ext_v2::{ExecutedBlockMeta, OrderedBlock, PipeExecLayerApi};
+use greth::reth_primitives::alloy_primitives::private::alloy_rlp::Decodable;
+use greth::reth_primitives::alloy_primitives::private::alloy_rlp::Encodable;
+use greth::reth_primitives::{
+    Address, TransactionSigned, TransactionSignedEcRecovered, Withdrawals, B256,
+};
+use greth::reth_provider::providers::BlockchainProvider2;
+use greth::reth_provider::{
+    AccountReader, BlockNumReader, BlockReaderIdExt, ChainSpecProvider, DatabaseProviderFactory,
+};
+use greth::reth_rpc_api::EngineEthApiClient;
+use greth::reth_rpc_types::BlockNumberOrTag;
+use greth::reth_transaction_pool::{PoolTransaction, TransactionPool};
 use std::io::Read;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 use tokio_stream::StreamExt;
-use tracing::info;
+use tracing::debug;
 use tracing::log::error;
-use web3::transports::Ipc;
-use web3::types::{BlockId, Transaction, TransactionId, H160};
-use web3::Web3;
 
 pub struct RethCli {
-    ipc: Web3<Ipc>,
     auth: AuthServerHandle,
-    pipe_api: Mutex<PipeExecLayerApi>,
+    pipe_api: PipeExecLayerApi,
     chain_id: u64,
+    provider: BlockchainProvider2<NodeTypesWithDBAdapter<EthereumNode, Arc<DatabaseEnv>>>,
+    txn_listener: Mutex<tokio::sync::mpsc::Receiver<greth::reth::primitives::TxHash>>,
+    pool: greth::reth_transaction_pool::Pool<
+        greth::reth_transaction_pool::TransactionValidationTaskExecutor<
+            greth::reth_transaction_pool::EthTransactionValidator<
+                BlockchainProvider2<NodeTypesWithDBAdapter<EthereumNode, Arc<DatabaseEnv>>>,
+                greth::reth_transaction_pool::EthPooledTransaction,
+            >,
+        >,
+        greth::reth_transaction_pool::CoinbaseTipOrdering<
+            greth::reth_transaction_pool::EthPooledTransaction,
+        >,
+        greth::reth_transaction_pool::blobstore::DiskFileBlobStore,
+    >,
 }
 
-pub fn covert_account(acc: H160) -> ExternalAccountAddress {
+pub fn covert_account(acc: greth::reth_primitives::Address) -> ExternalAccountAddress {
     let mut bytes = [0u8; 32];
-    bytes[12..].copy_from_slice(acc.as_bytes());
+    bytes[12..].copy_from_slice(acc.as_slice());
     ExternalAccountAddress::new(bytes)
 }
 
 impl RethCli {
-    fn create_payload_attributes(parent_beacon_block_root: B256, ts: u64) -> EthPayloadAttributes {
+    pub async fn new(args: ConsensusArgs) -> Self {
+        let chian_info = args.provider.chain_spec().chain;
+        let chain_id = match chian_info.into_kind() {
+            greth::reth_chainspec::ChainKind::Named(n) => n as u64,
+            greth::reth_chainspec::ChainKind::Id(id) => id,
+        };
+        RethCli {
+            auth: args.engine_api,
+            pipe_api: args.pipeline_api,
+            chain_id,
+            provider: args.provider,
+            txn_listener: Mutex::new(args.tx_listener),
+            pool: args.pool,
+        }
+    }
+
+    fn create_payload_attributes(
+        parent_beacon_block_root: greth::reth::primitives::B256,
+        ts: u64,
+    ) -> EthPayloadAttributes {
         EthPayloadAttributes {
             timestamp: ts,
-            prev_randao: B256::ZERO,
-            suggested_fee_recipient: Address::ZERO,
+            prev_randao: greth::reth::primitives::B256::ZERO,
+            suggested_fee_recipient: greth::reth_primitives::Address::ZERO,
             withdrawals: Some(Vec::new()),
             parent_beacon_block_root: Some(parent_beacon_block_root),
         }
-    }
-
-    pub async fn get_latest_block_hash(&self) -> Result<B256, String> {
-        let block = self.ipc.eth().block(BlockId::Number(web3::types::BlockNumber::Latest)).await;
-
-        match block {
-            Ok(Some(block)) => {
-                let hash = B256::from_slice(block.hash.unwrap().as_bytes());
-                Ok(hash)
-            }
-            _ => Err("Failed to get block".to_string()),
-        }
-    }
-
-    pub async fn new(ipc_url: &str, auth: AuthServerHandle, pipe_api: PipeExecLayerApi) -> Self {
-        let transport = web3::transports::Ipc::new(ipc_url).await.unwrap();
-        let ipc = web3::Web3::new(transport);
-        let chain_id = ipc.eth().chain_id().await.unwrap();
-        RethCli { ipc, auth, pipe_api: Mutex::new(pipe_api), chain_id: chain_id.as_u64() }
     }
 
     fn block_id_to_b256(block_id: ExternalBlockId) -> B256 {
         B256::new(block_id.0)
     }
 
-    fn construct_sig(txn: &Transaction) -> Signature {
-        let r = txn.r.unwrap();
-        let s = txn.s.unwrap();
-        let v = txn.v.unwrap();
-        let odd_y_parity = match v.as_u64() {
-            27 => false,
-            28 => true,
-            v => v % 2 == 1, 
-        };
-        let mut bytes = [0u8; 32];
-        r.to_big_endian(&mut bytes);
-        let r = U256::from_be_bytes(bytes);
-        let mut bytes = [0u8; 32];
-        s.to_big_endian(&mut bytes);
-        let s = U256::from_be_bytes(bytes);
-        Signature { r, s, odd_y_parity }
-    }
-
-    fn to_tx_kind(address: Option<H160>) -> TxKind {
-        match address {
-            Some(address) => TxKind::Call(Address::new(address.0)),
-            None => TxKind::Create,
-        }
-    }
-
-    fn convert_accest_list(access_list: Option<Vec<web3::types::AccessListItem>>) -> AccessList {
-        AccessList(
-            access_list
-                .unwrap_or_default()
-                .iter()
-                .map(|x| AccessListItem {
-                    address: Address::from_slice(x.address.as_bytes()),
-                    storage_keys: x
-                        .storage_keys
-                        .iter()
-                        .map(|x| B256::from_slice(x.as_bytes()))
-                        .collect(),
-                })
-                .collect(),
-        )
-    }
-
-    fn convert_to_reth_transaction(
-        tx: web3::types::Transaction,
-        chain_id: u64,
-    ) -> reth::primitives::Transaction {
-        match tx.transaction_type.map(|t| t.as_u64()) {
-            Some(0) => reth::primitives::Transaction::Legacy(TxLegacy {
-                chain_id: Some(chain_id),
-                nonce: tx.nonce.as_u64(),
-                gas_price: tx.gas_price.unwrap().as_u128(),
-                gas_limit: tx.gas.as_u128(),
-                to: Self::to_tx_kind(tx.to),
-                value: U256::from(tx.value.as_u128()),
-                input: Bytes::copy_from_slice(tx.input.0.as_slice()),
-            }),
-            Some(1) => reth::primitives::Transaction::Eip2930(TxEip2930 {
-                chain_id,
-                nonce: tx.nonce.as_u64(),
-                gas_price: tx.gas_price.unwrap().as_u128(),
-                gas_limit: tx.gas.as_u128(),
-                to: Self::to_tx_kind(tx.to),
-                value: U256::from(tx.value.as_u128()),
-                access_list: Self::convert_accest_list(tx.access_list),
-                input: Bytes::copy_from_slice(tx.input.0.as_slice()),
-            }),
-            Some(2) => reth::primitives::Transaction::Eip1559(TxEip1559 {
-                chain_id,
-                nonce: tx.nonce.as_u64(),
-                gas_limit: tx.gas.as_u128(),
-                to: Self::to_tx_kind(tx.to),
-                value: U256::from(tx.value.as_u128()),
-                access_list: Self::convert_accest_list(tx.access_list),
-                input: Bytes::copy_from_slice(tx.input.0.as_slice()),
-                max_fee_per_gas: tx.max_fee_per_gas.unwrap().as_u128(),
-                max_priority_fee_per_gas: tx.max_priority_fee_per_gas.unwrap().as_u128(),
-            }),
-            _ => panic!("Unknown transaction type {:?}", tx.transaction_type),
-        }
-    }
-
-    fn txn_to_signed(bytes: &[u8], chain_id: u64) -> (Address, TransactionSigned) {
-        let txn = serde_json::from_slice::<Transaction>(bytes).unwrap();
-        let address = txn.from.unwrap();
-        let address = Address::new(address.0);
-        let hash = txn.hash;
-        let hash = B256::from_slice(hash.as_bytes());
-        let signature = Self::construct_sig(&txn);
-        let transaction = Self::convert_to_reth_transaction(txn, chain_id);
-        info!("txn to signed {:?}", transaction);
-        info!("address {:?}", address);
-        (address, TransactionSigned { hash, signature, transaction })
+    fn txn_to_signed(bytes: &mut [u8], chain_id: u64) -> (Address, TransactionSigned) {
+        let txn = TransactionSignedEcRecovered::decode(&mut bytes.as_ref()).unwrap();
+        (txn.signer(), txn.into_signed())
     }
 
     pub async fn push_ordered_block(
         &self,
-        block: ExternalBlock,
-        parent_hash: B256,
-    ) -> Result<PayloadId, String> {
-        let pipe_api = self.pipe_api.lock().await;
-        let payload_attr =
-            Self::create_payload_attributes(parent_hash.into(), block.block_meta.usecs / 1000000);
-        let fcu_state = ForkchoiceState {
-            head_block_hash: parent_hash.into(),
-            safe_block_hash: parent_hash.into(),
-            finalized_block_hash: parent_hash.into(),
-        };
-        let engine_api = self.auth.http_client();
+        mut block: ExternalBlock,
+        parent_id: B256,
+    ) -> Result<(), String> {
+        debug!("push ordered block {:?} with parent id {}", block, parent_id);
+        let pipe_api = &self.pipe_api;
         let mut senders = vec![];
         let mut transactions = vec![];
         for (sender, txn) in
-            block.txns.iter().map(|txn| Self::txn_to_signed(&txn.bytes, self.chain_id))
+            block.txns.iter_mut().map(|txn| Self::txn_to_signed(&mut txn.bytes, self.chain_id))
         {
             senders.push(sender);
             transactions.push(txn);
         }
+
+        let randao = match block.block_meta.randomness {
+            Some(randao) => B256::from_slice(randao.0.as_ref()),
+            None => B256::ZERO,
+        };
+        // TODO: make zero make sense
         pipe_api.push_ordered_block(OrderedBlock {
-            block_id: Self::block_id_to_b256(block.block_meta.block_id),
-            parent_hash: parent_hash.into(),
+            parent_id,
+            id: B256::from_slice(block.block_meta.block_id.as_bytes()),
+            number: block.block_meta.block_number,
+            timestamp: block.block_meta.usecs / 1000000,
+            // TODO(gravity_jan): add reth coinbase
+            coinbase: Address::ZERO,
+            prev_randao: randao,
+            withdrawals: Withdrawals::new(Vec::new()),
             transactions,
             senders,
         });
-        let res = <HttpClient<AuthClientService<HttpBackend>> as EngineApiClient<EthEngineTypes>>::fork_choice_updated_v3(
-            &engine_api,
-            fcu_state, 
-            Some(payload_attr)
-        ).await;
-        match res {
-            Ok(res) => {
-                let payload_id = res.payload_id;
-                match payload_id {
-                    Some(payload_id) => Ok(payload_id),
-                    None => Err("Payload id not found".to_string()),
-                }
-            }
-            Err(_) => Err("Failed to push ordered block".to_string()),
-        }
+        Ok(())
     }
 
-    pub async fn process_payload_id(
-        &self,
-        block_id: B256,
-        payload_id: PayloadId,
-    ) -> Result<B256, ()> {
-        let mut pipe_api = self.pipe_api.lock().await;
-        let block_hash = pipe_api.pull_executed_block_hash(payload_id, block_id).await.unwrap();
+    pub async fn recv_compute_res(&self, block_id: B256) -> Result<B256, ()> {
+        debug!("recv compute res {:?}", block_id);
+        let pipe_api = &self.pipe_api;
+        let block_hash = pipe_api.pull_executed_block_hash(block_id).await.unwrap();
+        debug!("recv compute res done");
         Ok(block_hash)
     }
 
     pub async fn commit_block(
         &self,
-        parent_beacon_block_root: B256,
-        payload_id: PayloadId,
+        block_id: api_types::u256_define::BlockId,
         block_hash: B256,
     ) -> Result<(), String> {
-        let mut pipe_api = self.pipe_api.lock().await;
-        pipe_api.ready_to_get_payload(payload_id).await.unwrap();
-        let engine_api = self.auth.http_client();
-        let payload = <HttpClient<AuthClientService<HttpBackend>> as EngineApiClient<
-            EthEngineTypes,
-        >>::get_payload_v3(&engine_api, payload_id)
-        .await
-        .unwrap();
-        assert_eq!(payload.execution_payload.payload_inner.payload_inner.block_hash, block_hash);
-        pipe_api.ready_to_new_payload(block_hash.into()).await.unwrap();
-        let res = <HttpClient<AuthClientService<HttpBackend>> as EngineApiClient<EthEngineTypes>>::new_payload_v3(
-            &engine_api,
-            payload.execution_payload,
-            vec![],
-            parent_beacon_block_root,
-        ).await;
-        info!("payload status {:?}", res);
-        let res = <HttpClient<AuthClientService<HttpBackend>> as EngineApiClient<EthEngineTypes>>::fork_choice_updated_v3(
-            &engine_api,
-            ForkchoiceState {
-                head_block_hash: block_hash.into(),
-                safe_block_hash: block_hash.into(),
-                finalized_block_hash: block_hash.into(),
-            },
-            None
-        ).await;
-        match res {
-            Ok(_) => Ok(()),
-            Err(_) => Err("Failed to commit block".to_string()),
-        }
+        debug!("commit block {:?} with hash {:?}", block_id, block_hash);
+        let block_id = B256::from_slice(block_id.0.as_ref());
+        let pipe_api = &self.pipe_api;
+        pipe_api.commit_executed_block_hash(ExecutedBlockMeta { block_id, block_hash });
+        debug!("commit block done");
+        Ok(())
     }
 
     pub async fn process_pending_transactions(
         &self,
         buffer: Arc<Mutex<Vec<VerifiedTxnWithAccountSeqNum>>>,
     ) -> Result<(), String> {
-        let mut eth_sub =
-            self.ipc.eth_subscribe().subscribe_new_pending_transactions().await.unwrap();
-        info!("start process pending transactions");
-        while let Some(Ok(txn_hash)) = eth_sub.next().await {
-            info!("get txn hash {:?}", txn_hash);
-            let txn = self.ipc.eth().transaction(TransactionId::Hash(txn_hash)).await;
-            info!("get txn {:?}", txn);
-            if let Ok(Some(txn)) = txn {
-                let account = match txn.from {
-                    Some(account) => account,
-                    None => {
-                        error!("Transaction has no from account");
-                        continue;
-                    }
-                };
-                let accout_nonce = self.ipc.eth().transaction_count(account, None).await;
+        debug!("start process pending transactions");
+        let mut count = 0;
+        let mut total = 0;
+        let start_time = std::time::Instant::now();
+        let mut last_time = std::time::Instant::now();
+        let mut mut_txn_listener = self.txn_listener.lock().await;
+        while let Some(txn_hash) = mut_txn_listener.recv().await {
+            let txn = self.pool.get(&txn_hash).unwrap();
+            let before_recv = std::time::Instant::now();
+            let sender = txn.sender();
+            let nonce = txn.nonce();
+            let txn = txn.transaction.transaction();
+            let accout_nonce =
+                self.provider.basic_account(sender).unwrap().map(|x| x.nonce).unwrap_or(0);
+            let mut bytes = Vec::with_capacity(1024 * 4);
+            // txn.encode(&mut bytes);
+            txn.encode(&mut bytes);
 
-                match accout_nonce {
-                    Ok(accout_nonce) => {
-                        let mut buffer = buffer.lock().await;
-                        let bytes = serde_json::to_vec(&txn).unwrap();
-                        let committed_hash = TxnHash::new(simple_hash::hash_to_fixed_array(&bytes));
-                        let vtxn = VerifiedTxnWithAccountSeqNum {
-                            txn: VerifiedTxn {
-                                bytes,
-                                sender: covert_account(txn.from.unwrap()),
-                                sequence_number: txn.nonce.as_u64(),
-                                chain_id: ExternalChainId::new(0),
-                                committed_hash: committed_hash.into(),
-                            },
-                            account_seq_num: accout_nonce.as_u64(),
-                        };
-                        println!(
-                            "push txn nonce: {} acc_nonce: {}",
-                            txn.nonce, vtxn.account_seq_num
-                        );
-                        buffer.push(vtxn);
-                    }
-                    Err(e) => {
-                        error!("Failed to get nonce for account {:?} with {:?}", account, e);
-                    }
-                }
-            } else {
-                error!("Failed to get transaction {:?} {:?}", txn_hash, txn);
+            let vtxn = VerifiedTxnWithAccountSeqNum {
+                txn: VerifiedTxn {
+                    bytes,
+                    sender: covert_account(sender),
+                    sequence_number: nonce,
+                    chain_id: ExternalChainId::new(0),
+                    committed_hash: TxnHash::from_bytes(txn.hash().as_slice()).into(),
+                },
+                account_seq_num: accout_nonce,
+            };
+            {
+                count += 1;
+                let mut buffer = buffer.lock().await;
+                buffer.push(vtxn);
+            }
+            let after_ser = std::time::Instant::now();
+            debug!(
+                "push txn nonce: {} acc_nonce: {} recv_time {} serialize_time {}",
+                txn.transaction.nonce(),
+                accout_nonce,
+                before_recv.elapsed().as_micros(),
+                after_ser.elapsed().as_micros()
+            );
+            if last_time.elapsed().as_secs() > 1 {
+                debug!(
+                    "processed {} transactions in {}s with speed {}",
+                    count,
+                    last_time.elapsed().as_secs(),
+                    count as f64 / start_time.elapsed().as_secs_f64()
+                );
+                total += count;
+                count = 0;
+                last_time = std::time::Instant::now();
             }
         }
-        info!("end process pending transactions");
+
+        debug!("end process pending transactions");
         Ok(())
     }
-}
 
-#[derive(Serialize)]
-struct JsonRpcRequest<'a> {
-    jsonrpc: &'a str,
-    method: &'a str,
-    params: Vec<String>,
-    id: u64,
-}
+    pub async fn latest_block_number(&self) -> u64 {
+        match self.provider.header_by_number_or_tag(BlockNumberOrTag::Latest).unwrap() {
+            Some(header) => header.number, // The genesis block has a number of zero;
+            None => 0,
+        }
+    }
 
-#[derive(serde::Deserialize)]
-struct JsonRpcResponse<T> {
-    result: T,
+    pub async fn finalized_block_number(&self) -> u64 {
+        match self.provider.database_provider_ro().unwrap().last_block_number() {
+            Ok(block_number) => {
+                return block_number;
+            }
+            Err(e) => {
+                error!("finalized_block_number error {}", e);
+                return 0;
+            }
+        }
+    }
+
+    async fn recover_execution_blocks(&self, blocks: ExecutionBlocks) {}
+
+    pub fn get_blocks_by_range(
+        &self,
+        start_block_number: u64,
+        end_block_number: u64,
+    ) -> ExecutionBlocks {
+        let result = ExecutionBlocks {
+            latest_block_hash: todo!(),
+            latest_block_number: todo!(),
+            blocks: vec![],
+            latest_ts: todo!(),
+        };
+        for block_number in start_block_number..end_block_number {
+            match self.provider.block_by_number_or_tag(BlockNumberOrTag::Number(block_number)) {
+                Ok(block) => {
+                    assert!(block.is_some());
+                    let block = block.unwrap();
+                    if block_number == end_block_number - 1 {
+                        result.latest_block_hash = *block.hash_slow();
+                        result.latest_block_number = block_number;
+                        result.latest_ts = block.timestamp;
+                    }
+                    result.blocks.push(bincode::serialize(&block).unwrap());
+                }
+                Err(e) => panic!("get_blocks_by_range error {}", e),
+            }
+        }
+        result
+    }
 }
