--- bin/pethv2/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "pethv2"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license.workspace = true
-publish.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-
-
-[dependencies]
-futures-util = "0.3.30"
-clap = { version = "4.5.17", features = ["derive", "env"] }
-futures = "0.3.30"
-dirs = "5.0.1"
-tokio = "1.40.0"
-tokio-stream = "0.1.16"
-eyre = "0.6.12"
-tracing = "0.1.40"
-tempfile = "3.12.0"
-jsonrpsee = "0.24"
-hex = "0.4.3"
-serde_json = "1.0.128"
-bincode = "1.3"
-time = "0.3.36"
-anyhow = "1.0.87"
-greth = { git = "https://github.com/Galxe/grevm-reth", branch = "add_lib" }
-reqwest = "0.12.9"
-
-async-trait.workspace = true
-api.workspace = true
-api-types.workspace = true
-
