--- bin/pethv2/src/consensus/mod.rs
@@ -1,3 +1,0 @@
-pub mod aptos;
-pub mod mempool;
-pub mod mock;
